refactor(auth): align the auth route files hierarchy with the auth routes
