+++ src/routes/auth/email/index.ts
