+++ src/routes/auth/index.ts
@@ -1,6 +1,8 @@
 export * from './mfa'
 export * from './user'
 export * from './token'
+export * from './email'
+export * from './password'
 
 export { default as getJwks } from './jwks'
 export { default as loginUser } from './login'
