+++ src/routes/auth/password/forgot.ts
