+++ src/routes/auth/password/index.ts
