+++ src/routes/auth/user/index.ts
@@ -1,5 +1,2 @@
-export * from './email'
-export * from './password'
-
 export { default as deleteUser } from './delete'
 export { default as activateUser } from './activate'
